+++ mycore/settings.py
@@ -56,6 +56,7 @@ INSTALLED_APPS = [
     'mywallet',
     'mynotice',
     'myfundingprojects',
+    'mylikelists',
 ]
 
 MIDDLEWARE = [
