+++ myfundingprojects/views.py
@@ -1,5 +1,5 @@
-from mydatabase.models import UserDatas,FundingProjects,LikeLists,FundingShares
-from myfundingprojects.serializers import FundingProjectsSerializer,UserLikeListsSerializer,UserFundingSharesSerializer
+from mydatabase.models import UserDatas,FundingProjects,FundingShares
+from myfundingprojects.serializers import FundingProjectsSerializer,UserFundingSharesSerializer
 from rest_framework import viewsets
 from rest_framework import status
 from rest_framework.response import Response
@@ -31,26 +31,7 @@ class FundingProjectsViewSet(viewsets.ModelViewSet):
         except:
                 return Response(status=status.HTTP_400_BAD_REQUEST) 
 
-class UserLikeListsViewSet(viewsets.ModelViewSet):
-    permission_classes = [IsAuthenticated,]
 
-    serializer_class = UserLikeListsSerializer
-    queryset = LikeLists.objects.all()
-
-    def get_queryset(self):                                            # added string
-        return super().get_queryset().filter(userData=UserDatas.objects.get(id=self.request.user.id))
-    
-    def create(self, request, *args, **kwargs):
-        userLikeListsdata=request.data
-        
-        try:
-                new_userLikeLists = LikeLists.objects.create(
-                userData=UserDatas.objects.get(id=self.request.user.id),
-                fundingProject=FundingProjects.objects.get(id=userLikeListsdata['fundingProject']) )
-                new_userLikeLists.save()
-                return Response(status=status.HTTP_201_CREATED) 
-        except:
-                return Response(status=status.HTTP_400_BAD_REQUEST) 
 
 class UserFundingSharesViewSet(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticated,]
