+++ mylikelists/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework.serializers import ModelSerializer
+from mydatabase.models import LikeLists
+
+
+class UserLikeListsSerializer(ModelSerializer):
+    
+    class Meta:
+        model=LikeLists
+        fields = ['id','fundingProject']
